Make listen actually work, and be honest about what it does.
Note that this is a backwards-incompatible change, in so far as the body of the message is returned in :body as opposed to :message, and there is no :person field with the name of the user. I think I'd like to use the user_id to look up the user from the users in the room and return a hash with the user information (or at least :id and :name), but that will be a separate commit.

Also, note that there is a bug in ruby versions before 1.8.7-p249 that prevents listen from working, as Yajl::HttpStream.get returns immediately. See http://github.com/brianmario/yajl-ruby/issues/closed#issue/14